move write offset length to new expression
