+++ packages/core/src/abap/2_statements/expressions/index.ts
@@ -173,3 +173,4 @@ export * from "./type";
 export * from "./value_body";
 export * from "./value";
 export * from "./with_name";
+export * from "./write_offset_length";
